autonama.data/requirements.txt
# Task Queue & Processing
celery>=5.5.3
redis>=5.2.0
flower>=2.0.1
kombu>=5.4.2

# Database Connections
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
# duckdb>=1.3.1  # REMOVED: No longer needed - all calculations done locally

# Data Processing & Analysis
pandas>=2.2.0
numpy>=1.24.0,<2.0  # Pin to NumPy 1.x for compatibility
scipy>=1.11.0  # Required for VectorBTPro optimization
# scikit-learn>=1.6.0
# statsmodels>=0.14.4

# Trading & Market Data
ccxt>=4.4.0
yfinance>=0.2.48
python-binance>=1.0.19
# alpha-vantage>=2.3.1
# polygon-api-client>=1.15.0

# Technical Analysis
# ta-lib>=0.4.32
pandas-ta>=0.3.14b0
# finta>=1.3

# Optimization & ML
optuna>=4.1.0
# hyperopt>=0.2.7
# bayesian-optimization>=1.5.1
# xgboost>=2.1.3
# lightgbm>=4.5.0

# Backtesting & Strategy
# backtrader>=1.9.78.123
# zipline-reloaded>=3.0.4

# Utilities & Configuration
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
loguru>=0.7.2
click>=8.1.7

# Logging & ELK Integration
python-json-logger>=2.0.7
structlog>=24.4.0

# HTTP & API
httpx>=0.28.0
requests>=2.32.0
aiohttp>=3.11.0
websocket-client>=1.6.0

# Development & Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.10.0
isort>=5.13.0
flake8>=7.1.0
mypy>=1.13.0
psutil>=5.9.0
beautifulsoup4>=4.12.0
# pyarrow>=12.0.0  # Optional - only needed for local parquet operations 